requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyYAML==6.0.2
llama-cpp-python==0.3.8
